% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_indic_sign.R
\name{add_indic_sign}
\alias{add_indic_sign}
\title{Add indic in dataset of simulated fitness}
\usage{
add_indic_sign(dataset_sim = sim)
}
\arguments{
\item{dataset_sim}{dataset with simulated data from the simul_fitnessdata_normal()}
}
\value{
the same dataset with new columns
}
\description{
Add indic if each simul is significant for genetic and non genetic test
}
\examples{
simul_fitnessdata_normal(dataset_sim = sim)
}
