% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sim_number_SA.R
\name{add_sim_number_SA}
\alias{add_sim_number_SA}
\title{Add index for the number of simul for each SA value}
\usage{
add_sim_number_SA(dataset_sim = sim)
}
\arguments{
\item{dataset_sim}{dataset with simulated data from the simul_fitnessdata_normal()}
}
\value{
the same dataset with new columns
}
\description{
Add index for the number of simul for each SA value
}
\examples{
add_sim_number_SA(dataset_sim = sim)
}
