% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_all.R
\docType{data}
\name{datablank}
\alias{datablank}
\title{Legend for plot}
\format{
An object of class \code{data.frame} with 3 rows and 2 columns.
}
\usage{
datablank
}
\value{

}
\description{
Create a legend for most of the plot
}
\examples{
PLOT_MAIN_Text<-ggplot + legend_all
}
\keyword{datasets}
