% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_rtp.R
\docType{data}
\name{datablank_rtp}
\alias{datablank_rtp}
\title{Legend for plot}
\format{
An object of class \code{data.frame} with 3 rows and 3 columns.
}
\usage{
datablank_rtp
}
\value{

}
\description{
Create a legend for most of the plot
}
\examples{
PLOT_MAIN_Text<-ggplot + legend_all
}
\keyword{datasets}
