% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Fitness Data Import Function}
\usage{
import_data(
  dataset = "DATACOMPLET_PERF.csv",
  trait = "performance",
  remove_testenvt = c("Grape", "GF"),
  remove_pop = "WT3",
  remove_rate = NA
)
}
\arguments{
\item{dataset}{Fitness dataset}

\item{trait}{can be only "preference" of "performance"}

\item{remove_testenvt}{remove all data measured on this test_environment (could be a vector with several test environment)}

\item{remove_pop}{remove all data of this population}

\item{remove_rate}{if TRUE remove data where number of adults is higher than the number of eggs; if NA doesnt change; if "Replace" replace Rate by 1}
}
\value{

}
\description{
Reads a fitness data file in table format and subset it to return the corrct dataframe
}
\examples{
data_PERF <- import_data(dataset = "DATACOMPLET_PERF.csv", trait = "performance",
remove_testenvt = c("Grape","GF"), remove_pop = "WT3", remove_rate = NA)
}
