% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_acp.R
\name{plot_acp}
\alias{plot_acp}
\title{Plot pairwise residuals pop effect}
\usage{
plot_acp(
  dataset = data_PERF_Rate,
  formula = "asin(sqrt(Rate)) ~ Test_environment + log(Nb_eggs)",
  gen = "G2",
  titleplot = "Offspring performance in G0"
)
}
\arguments{
\item{dataset}{name of the dataset (tibble format)}

\item{formula}{formula of the model to use to compute the residuals}

\item{gen}{can be G0 or G2}
}
\value{
plot
}
\description{
Create a bivariate plot with residuals from common garden experiment
}
\examples{
plot_acp(dataset = data_PERF_Rate, formula = modelformula, gen = "G2")
}
