% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation_meanresiduals.R
\name{plot_correlation_meanresiduals}
\alias{plot_correlation_meanresiduals}
\title{Plot the relationship between a trait measured using an environmental factor with two levels}
\usage{
plot_correlation_meanresiduals(
  dataset = data,
  formula = "log(Nb_adults+1) ~",
  envfactor = "Env",
  original_environment = NULL,
  additional_factor = NULL,
  col_original_environment = NULL,
  grp_cols = c("Pop", "Fruit", "Env"),
  facetwrap = TRUE,
  printnames = FALSE,
  printsamplesize = FALSE,
  labelsize = 1,
  xaxis_labelprint = NULL,
  yaxis_labelprint = NULL,
  xlim = NULL,
  ylim = NULL,
  fixedxylim = TRUE,
  bisector = TRUE,
  errorbars = TRUE
)
}
\arguments{
\item{dataset}{name of the dataset (tibble format)}

\item{formula}{formula of the model to use to compute the residuals}

\item{envfactor}{environmental factor with two levels}

\item{original_environment}{name of the column including the environment of origin of each population}

\item{additional_factor}{name of the column of an additional factor whose level will be represented with different colors or on different panel (facetwrap=TRUE) on the final graph}

\item{col_original_environment}{vector of color for the different levels of the factor "original_environment"}

\item{grp_cols}{vector with the names of factors that should be used to compute the mean of the trait (first/last element: name of the column with the population and test environment information)}

\item{facetwrap}{split additional factor into different panels}

\item{printnames}{(logical) use the name of the first factor listed in grp_cols as labels (default=FALSE)}

\item{printsamplesize}{(logical) print the sample size (x, y) as labels (default=FALSE)}

\item{labelsize}{size of the labels if printed}

\item{xaxis_labelprint}{label of the x axis}

\item{yaxis_labelprint}{label of the y axis}

\item{xlim}{scale for x-axis}

\item{ylim}{scale for y-axis}

\item{fixedxylim}{(logical) should the limits for the scale of x-axis and y-axis be computed a the range of the trait across all test_environments (default=FALSE)}

\item{bisector}{(logical) print a bisector line through the origin (default=FALSE)}
}
\description{
Create a bivariate plot showing the relationship between a trait measured using an environmental factor with two levels
}
