% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Correlationtraits_residuals.R
\name{plot_CorrelationTraits_residuals}
\alias{plot_CorrelationTraits_residuals}
\title{Plot for relationship between rate and oviposition}
\usage{
plot_CorrelationTraits_residuals(
  gen = "G2",
  fruit = "All",
  dataset1 = data_PERF_Rate,
  dataset2 = data_PREF_three,
  formula1 = "asin(sqrt(Rate)) ~ Test_environment + Population + log(Nb_eggs)",
  formula2 = "log(Nb_eggs+1) ~ Test_environment+ Population + BoxID",
  axis_label_trait1 = "Offspring performance in XX",
  axis_label_trait2 = "Oviposition preference in XX",
  subscript_label = "preference/performance",
  printcor = TRUE
)
}
\arguments{
\item{gen}{the generation of dataset but can be: "G0", "G2" or  "Both"}

\item{fruit}{could be "Cherry" "Strawberry" "Blackberry" or "All"}

\item{dataset1}{name of the dataset (tibble format) for the trait 1 = x-axis}

\item{dataset2}{name of the dataset (tibble format) for the trait 2 = y-axis}

\item{formula1}{formula of the model to use to compute the residuals of the trait 1 = x-axis}

\item{formula2}{formula of the model to use to compute the residuals of the trait 2 = y-axis}

\item{axis_label_trait1}{label of the x axis}

\item{axis_label_trait2}{label of the y axis}

\item{printcor}{print correlation between the two variables}

\item{subscript}{subscript used when printing the correlation on the final graph}
}
\value{
plot of the residuals of the trait on the three environement
}
\description{
Create plot for relation between two traits
}
\examples{
plot_CorrelationTraits_residuals(gen = "G2", fruit = "All", dataset1 = data_PERF_Rate, dataset2 = data_PREF_three, formula1 = "asin(sqrt(Rate)) ~ Test_environment + log(Nb_eggs)", formula2 = "log(Nb_eggs+1) ~ Test_environment + BoxID", axis_label_trait1 = "Offspring performance in XX", axis_label_trait2 = "Oviposition preference in XXX", subscript_label = "preference/performance", printcor=TRUE)
}
