% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Genetic_Nongenetic_residuals.R
\name{plot_Genetic_Nongenetic_residuals}
\alias{plot_Genetic_Nongenetic_residuals}
\title{Plot for reciprocal transplant experiment}
\usage{
plot_Genetic_Nongenetic_residuals(
  dataset = data_PERF_Rate,
  trait = "Rate",
  effect = "Genetic"
)
}
\arguments{
\item{trait}{the trait}

\item{effect}{can be "Genetic" or "Non-genetic"}
}
\value{
plot of the residuals of the trait on the three environement
}
\description{
Create plot for reciprocal transplant experiment
}
\examples{
plot_RTP_residuals(dataset = data_PERF, trait = "Nb_eggs", effect = "Genetic")
}
