% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Pairwise_meanresiduals_popeffect.R,
%   R/plot_Pairwise_meanresiduals_popeffect_Alltogether.R
\name{plot_pairwise_meanresiduals_popeffect}
\alias{plot_pairwise_meanresiduals_popeffect}
\title{Plot pairwise residuals}
\usage{
plot_pairwise_meanresiduals_popeffect(
  dataset = data_PERF_Rate,
  formula = "asin(sqrt(Rate)) ~ Test_environment + log(Nb_eggs)",
 
    formula_Blanquart = "asin(sqrt(Rate)) ~  Test_environment + Population + SA + log(Nb_eggs) + Test_environment:Original_environment",
  test_environment = "Test_environment",
  original_environment = "Original_environment",
  gen = "G2",
  subscript = NA,
  grp_cols = c("Population", "Original_environment", "Test_environment"),
  coltest_envlevels = c("#301934", "#BC3C6D", "#3FAA96"),
  xaxis_labelprint = "Offspring performance\\nin sympatry",
  yaxis_labelprint = "Offspring performance\\nin allopatry",
  xlim = NULL,
  ylim = NULL,
  fixedxylim = TRUE,
  bisector = FALSE,
  printcor = FALSE,
  printSA = FALSE,
  test_Blanquart = TRUE
)

plot_pairwise_meanresiduals_popeffect(
  dataset = data_PERF_Rate,
  formula = "asin(sqrt(Rate)) ~ Test_environment + log(Nb_eggs)",
 
    formula_Blanquart = "asin(sqrt(Rate)) ~  Test_environment + Population + SA + log(Nb_eggs) + Test_environment:Original_environment",
  test_environment = "Test_environment",
  original_environment = "Original_environment",
  gen = "G2",
  subscript = NA,
  grp_cols = c("Population", "Original_environment", "Test_environment"),
  coltest_envlevels = c("#301934", "#BC3C6D", "#3FAA96"),
  xaxis_labelprint = "Offspring performance\\nin sympatry",
  yaxis_labelprint = "Offspring performance\\nin allopatry",
  xlim = NULL,
  ylim = NULL,
  fixedxylim = TRUE,
  bisector = FALSE,
  printcor = FALSE,
  printSA = FALSE,
  test_Blanquart = TRUE
)
}
\arguments{
\item{dataset}{name of the dataset (tibble format)}

\item{formula}{formula of the model to use to compute the residuals}

\item{formula_Blanquart}{formula of the model to estimate Blanquart Local Adaptation}

\item{test_environment}{name of the column including the test environment}

\item{original_environment}{name of the column including the environment of origin of each population}

\item{gen}{can be G0 or G2}

\item{subscript}{subscript used when printing the correlation on the final graph}

\item{grp_cols}{vector with the names of factors that should be used to compute the mean of the trait (first/last element: name of the column with the population and test environment information)}

\item{coltest_envlevels}{vector of color for the different levels of the factor "test_environment"}

\item{xaxis_labelprint}{label of the x axis}

\item{yaxis_labelprint}{label of the y axis}

\item{xlim}{scale for x-axis}

\item{ylim}{scale for y-axis}

\item{fixedxylim}{(logical) should the limits for the scale of x-axis and y-axis be computed a the range of the trait across all test_environments (default=FALSE)}

\item{bisector}{(logical) print a bisector line through the origin (default=FALSE)}

\item{printcor}{print correlation between the two variables}

\item{printSA}{print SA}

\item{test_Blanquart}{print the results of the Blanquart's SA test}

\item{errorbars}{print bars for standard errors}

\item{fruit1}{fruit for y-axis}

\item{fruit2}{fruit for x-axis}

\item{onlyfocalpop}{(logical) use only populations sampled from fruit1 or fruit2  (default=TRUE)}
}
\value{
plot

plot
}
\description{
Create a bivariate plot with residuals from common garden experiment

Create a bivariate plot with residuals from common garden experiment
}
\examples{
plot_pairwise_meanresiduals(dataset = data_PERF_Rate, formula=modelformula, fruit1 = "Cherry", fruit2 = "Blackberry", 
grp_cols=c("City", "Generation", "Original_environment", "Test_environment") , test_environment="Test_environment", original_environment="Original_environment", additional_factor="Generation", coltest_envlevels = c("#301934","#BC3C6D", "#3FAA96"), fixedxylim = TRUE, bisector = TRUE)
plot_pairwise_meanresiduals(dataset = data_PERF_Rate, formula=modelformula, fruit1 = "Cherry", fruit2 = "Blackberry", 
grp_cols=c("City", "Generation", "Original_environment", "Test_environment") , test_environment="Test_environment", original_environment="Original_environment", additional_factor="Generation", coltest_envlevels = c("#301934","#BC3C6D", "#3FAA96"), fixedxylim = TRUE, bisector = TRUE)
}
