% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PairwisePOP_residuals.R
\name{plot_PairwisePOP_residuals}
\alias{plot_PairwisePOP_residuals}
\title{Plot for reciprocal transplant experiment}
\usage{
plot_PairwisePOP_residuals(
  dataset = data_PERF_Rate,
  trait = "Rate",
  gen = "G2",
  fruit1 = "Cherry",
  fruit2 = "Blackberry"
)
}
\arguments{
\item{trait}{the trait}

\item{gen}{the generation of dataset but can be Both}

\item{fruit1}{fruit for y-axis}

\item{fruit2}{fruit for x-axis}
}
\value{
plot of the residuals of the trait on the three environement
}
\description{
Create plot for reciprocal transplant experiment
}
\examples{
plot_PairwisePOP_residuals(dataset = data_PERF, trait = "Nb_eggs", gen = "Both",fruit1 = "Cherry",fruit2 = "Blackberry")
}
