% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pop_homeaway.R
\name{plot_pop_homeaway}
\alias{plot_pop_homeaway}
\title{Plot pairwise residuals pop effect}
\usage{
plot_pop_homeaway(
  dataset = data_PREF_three,
  trait = "Nb_eggs",
  test_environment = "Test_environment",
  original_environment = "Original_environment",
  gen = "G2",
  grp_cols = c("Population", "Original_environment", "Test_environment"),
  coltest_envlevels = c("#301934", "#BC3C6D", "#3FAA96"),
  yaxis_labelprint = "Offspring performance\\nin allopatry"
)
}
\arguments{
\item{dataset}{name of the dataset (tibble format)}

\item{test_environment}{name of the column including the test environment}

\item{original_environment}{name of the column including the environment of origin of each population}

\item{gen}{can be G0 or G2}

\item{grp_cols}{vector with the names of factors that should be used to compute the mean of the trait (first/last element: name of the column with the population and test environment information)}

\item{coltest_envlevels}{vector of color for the different levels of the factor "test_environment"}

\item{yaxis_labelprint}{label of the y axis}
}
\value{
plot
}
\description{
Create a bivariate plot with residuals from common garden experiment
}
\examples{
plot_realdata(dataset = data_PERF_Rate, formula=modelformula, fruit1 = "Cherry", fruit2 = "Blackberry", 
grp_cols=c("City", "Generation", "Original_environment", "Test_environment") , test_environment="Test_environment", original_environment="Original_environment", additional_factor="Generation", coltest_envlevels = c("#301934","#BC3C6D", "#3FAA96"), fixedxylim = TRUE)
}
