% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RelationTraits_residuals.R
\name{plot_RelationTraits_residuals}
\alias{plot_RelationTraits_residuals}
\title{Plot for relationship between rate and oviposition}
\usage{
plot_RelationTraits_residuals(
  gen = "G2",
  fruit = "Blackberry",
  trait2 = "Preference"
)
}
\arguments{
\item{gen}{the generation of dataset but can be: "G0", "G2" or  "Both"}

\item{fruit}{could be "Cherry" "Strawberry" or "Blackberry"}

\item{trait2}{Could be "Stimulation" or "Preference" (trait 1 is Egg-to-adult viabilty)}
}
\value{
plot of the residuals of the trait on the three environement
}
\description{
Create plot for relation between egg to adult and oviposition stimulation
}
\examples{
plot_PairwisePOP_residuals(dataset = data_PERF, trait = "Nb_eggs", gen = "Both",fruit1 = "Cherry",fruit2 = "Blackberry")
}
