% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RTP_realdata.R
\name{plot_RTP_realdata}
\alias{plot_RTP_realdata}
\title{Plot for reciprocal transplant experiment with REAL DATA}
\usage{
plot_RTP_realdata(dataset = data_PREF_three, trait = "Nb_eggs", gen = "G0")
}
\arguments{
\item{trait}{the trait}

\item{gen}{the generation of dataset}
}
\value{
plot of the real data of the trait on the three environement
}
\description{
Create plot for reciprocal transplant experiment with real data
}
\examples{
plot_RTP_residuals(dataset = data_PERF, trait = "Nb_eggs", gen = "G0")
}
