% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RTP_residuals.R
\name{plot_RTP_residuals}
\alias{plot_RTP_residuals}
\title{Plot for reciprocal transplant experiment}
\usage{
plot_RTP_residuals(dataset = data_PERF, trait = "Nb_eggs", gen = "G0")
}
\arguments{
\item{trait}{the trait}

\item{gen}{the generation of dataset}
}
\value{
plot of the residuals of the trait on the three environement
}
\description{
Create plot for reciprocal transplant experiment
}
\examples{
plot_RTP_residuals(dataset = data_PERF, trait = "Nb_eggs", gen = "G0")
}
