% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_allprop_figures.R
\name{save_allprop_figures}
\alias{save_allprop_figures}
\title{Save all the dataset and merge them}
\usage{
save_allprop_figures(
  data_prop_to_add = data_prop,
  distribution = "normal",
  generation = "genetic"
)
}
\arguments{
\item{data_prop}{dataset with simulated data from the simul_fitnessdata_normal()}
}
\value{
the same dataset with new columns
}
\description{
Create dataset with all the simuls
}
\examples{
simul_fitnessdata_normal(dataset_sim = sim)
}
