% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_sample_simul.R
\name{select_sample_simul}
\alias{select_sample_simul}
\title{Select a sample of simulations}
\usage{
select_sample_simul(
  dataset_sim = sim,
  type_SA = "genetic",
  nb_simul_per_SA = 100,
  minimum_SA = "0"
)
}
\arguments{
\item{dataset_sim}{dataset with simulated data after index with index the add_sim_number_SA()}

\item{type_SA}{Select on SA genetic or SA non-genetic}

\item{nb_simul_per_SA}{The number of simulations kept for each SA value}

\item{minimum_SA}{To select or not on the minimum values of SA
If "0": No select dataset with val_SA_available_MIN. The dataset will start with SA=0 (val_SA_available_MIN=0) even if less than nb_simul_per_SA for SA=0.
If "select": Select dataset with val_SA_available_MIN. The lowest SA will be the lowest SA with at least nb_simul_per_SA simuls.}
}
\value{
the same dataset with new columns
}
\description{
Select the same number of simul for each SA value (Avoid over-representing some SA values)
}
\examples{
select_sample_simul(dataset_sim = sim)
}
