% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_fitnessdata.R
\name{simul_fitnessdata}
\alias{simul_fitnessdata}
\title{Simulate fitness and calculate SA}
\usage{
simul_fitnessdata(
  distrib = "normal",
  design = "balanced",
  seed = 1,
  nfruit = 3,
  nhab = 3,
  npop_per_fruit = 5,
  nrep = 3,
  ntrial = 20,
  sdbox = NA,
  sdpop = 0.5,
  sdfruithab = 1,
  sdfruithab_ng = 1,
  rho = -1/(nhab - 1),
  rho_ng = 0,
  sigma = 0.5,
  disp_design = 1e-04
)
}
\arguments{
\item{distrib}{Distribution of data: "normal", "poisson" or "binomial"}

\item{design}{If design = "unbalanced" -> sample dataset}

\item{seed}{Seed used for simulation}

\item{nfruit}{Number of selective fruits}

\item{nhab}{Number of habitat}

\item{npop_per_fruit}{Number of populations per fruit}

\item{nrep}{Number of replicates per populationxhabitat}

\item{ntrial}{Number of trial of Bernouilli (only for Logit distribution)}

\item{sdpop}{Standard error for the variance across populations}

\item{sdfruithab}{Standard error for the variance of the genetic interaction PopxHabitat}

\item{sdfruithab_ng}{Standard error for the variance of the non-genetic interaction PopxHabitat}

\item{rho}{Genetic covariance (from which emerges SA genetic)}

\item{rho_ng}{Non genetic covariance (from which emerges SA non genetic)}

\item{sigma}{Residual variance across replicates}
}
\value{
vector seed, SA_Gen_True, SA_NonGen_True, SAcoef, rho, rho_ng, Fratio_Gen,
pvalue_Gen, Fratio_NonGen, pvalue_NonGen, Fratio_Gen_aov, pvalue_Gen_aov, Fratio_NonGen_aov, pvalue_NonGen_aov
}
\description{
Simulate fitness data
}
\examples{
simul_fitnessdata_normal(distrib = "normal", design = "balanced", seed = 1, nfruit = 3, nhab = 3, npop_per_fruit = 3,
nrep = 1, sdpop = 0,  sdfruithab = 1, sdfruithab_ng = 1, rho=-1/(nhab-1), rho_ng=-1/(nhab-1)
}
