% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_fitnessdata_poisson.R
\name{simul_fitnessdata_poisson}
\alias{simul_fitnessdata_poisson}
\title{Simulate fitness and calculate SA with poisson distribution}
\usage{
simul_fitnessdata_poisson(
  seed = 1,
  npop = 3,
  nhab = 3,
  nrep = 5,
  sdpophab = 1,
  rho = -1/(nhab - 1),
  rho_ng = -1/(nhab - 1),
  sdpophab_ng = 1,
  sigma = 0.5,
  constant = 0
)
}
\arguments{
\item{seed}{Seed used for simulation}

\item{npop}{Number of populations}

\item{nhab}{Number of habitat}

\item{nrep}{Number of replicates per populationxhabitat}

\item{sdpophab}{Standard error for the variance of the genetic interaction PopxHabitat}

\item{rho}{Genetic covariance (from which emerges SA genetic)}

\item{rho_ng}{Non genetic covariance (from which emerges SA non genetic)}

\item{sdpophab_ng}{Standard error for the variance of the non-genetic interaction PopxHabitat}

\item{sigma}{Residual variance across replicates}

\item{constant}{Should we add a constant factor?}
}
\value{
vector seed, SA_Gen_True, SA_NonGen_True, SAcoef, rho, rho_ng, Fratio_Gen,
pvalue_Gen, Fratio_NonGen, pvalue_NonGen, Fratio_Gen_aov, pvalue_Gen_aov, Fratio_NonGen_aov, pvalue_NonGen_aov
}
\description{
Simulate fitness data: poisson distribution
}
\examples{
simul_fitnessdata_poisson(seed = 1, npop = 3, nhab = 3, nrep = 5, sdpophab = 1, sdpophab_ng = 1, 
sigma = 0.5, rho=-1/(nhab-1), rho_ng=-1/(nhab-1), sdpophab_ng = 1, sigma = 0.5, constant = 0)
}
