% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_fitnessdata_unbalanced.R
\name{simul_fitnessdata_unbalanced}
\alias{simul_fitnessdata_unbalanced}
\title{Simulate fitness and calculate SA}
\usage{
simul_fitnessdata_unbalanced(
  distrib = "normal",
  unbalanced_dataset = data_PERF,
  seed = 1,
  sdpop = 0.5,
  sdfruithab = 1,
  sdfruithab_ng = 1,
  rho = -1/(nhab - 1),
  rho_ng = 0,
  sigma = 0.5
)
}
\arguments{
\item{distrib}{Distribution of data: "normal", "poisson" or "binomial"}

\item{unbalanced_dataset}{Unbalanced dataset}

\item{seed}{Seed used for simulation}

\item{sdpop}{Standard error for the variance across populations}

\item{sdfruithab}{Standard error for the variance of the genetic interaction PopxHabitat}

\item{sdfruithab_ng}{Standard error for the variance of the non-genetic interaction PopxHabitat}

\item{rho}{Genetic covariance (from which emerges SA genetic)}

\item{rho_ng}{Non genetic covariance (from which emerges SA non genetic)}

\item{sigma}{Residual variance across replicates}
}
\value{
vector seed, SA_Gen_True, SA_NonGen_True, SAcoef, rho, rho_ng, Fratio_Gen,
pvalue_Gen, Fratio_NonGen, pvalue_NonGen, Fratio_Gen_aov, pvalue_Gen_aov, Fratio_NonGen_aov, pvalue_NonGen_aov
}
\description{
Simulate fitness data
}
\examples{
simul_fitnessdata_unbalanced(distrib = "normal", unbalanced_dataset = data_PERF, seed = 1, sdpop = 0.5, 
sdfruithab = 1, sdfruithab_ng = 1, rho = -1/(nhab-1), rho_ng = 0, sigma = 0.)
}
