% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_powertest_nongenet.R
\name{simul_powertest_nongenet}
\alias{simul_powertest_nongenet}
\title{Power test non genetic local adaptation}
\usage{
simul_powertest_nongenet(seed = 1, SAng = 0.32)
}
\arguments{
\item{seed}{Seed used for simulation}

\item{SAng}{Estimate of the  non genetic SA effect}
}
\value{
vector: seed, SAng, SAGen, SANonGen, SAGen_Est, SANonGen_Est, Fratio_Gen,
pvalue_Gen, Fratio_NonGen, pvalue_NonGen, IndicGen, IndicNonGen
}
\description{
Simulation to estimate the power of detection non genetic local adaptation
}
\examples{
simul_powertest_nongenet(seed = 1, SAng = 0.32)
}
