% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulLA.R
\name{simulateLA}
\alias{simulateLA}
\title{Simulate Local Adaptation Function}
\usage{
simulateLA(SA)
}
\arguments{
\item{SA}{fitness advantage in sympatry}
}
\value{

}
\description{
Simulate data with local adaptation
}
\examples{
ddply(data_logchangeNb_eggsEmergence_rate, .(Generation, Fruit_s), computecorrelation)
}
